add keys hook
