+++ .vscode/settings.json
@@ -4,6 +4,7 @@
         "string": "c",
         "string_view": "c",
         "ranges": "c",
-        "span": "c"
+        "span": "c",
+        "cub3d.h": "c"
     }
 }
